Initial Commit
Ionic 5|4 Native In App Purchase Tutorial with Example Application

In this Ionic 5/4 tutorial, we'll integrate In-App Purchase
functionality in Android and iOS applications using Cordova and Ionic
Native plugins.

The In-App Purchase feature allows a user to buy from a
list of available products in the Application store. Use can make
purchases from a list of items available make payments and also restore
the purchase.

Link:
https://www.freakyjolly.com/ionic-native-in-app-purchase-tutorial-with-example-application/